Move inference engines to neuralogic.inference
