+++ neuralogic/inference/__init__.py
@@ -1,0 +1,2 @@
+from neuralogic.inference.evaluation_inference_engine import EvaluationInferenceEngine
+from neuralogic.inference.inference_engine import InferenceEngine
